#91: Refactor pattern base class.
