+++ src/Plugin/PatternInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Drupal\ui_patterns\Plugin;
+
+/**
+ * Interface PatternInterface.
+ *
+ * @package Drupal\ui_patterns
+ */
+interface PatternInterface {
+
+}
