--- src/UiPatternInterface.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Drupal\ui_patterns;
-
-/**
- * Interface UiPatternInterface.
- *
- * @package Drupal\ui_patterns
- */
-interface UiPatternInterface {
-
-}
