+++ ui_patterns.module
@@ -11,7 +11,7 @@ use Drupal\ui_patterns\UiPatterns;
  * Implements hook_theme().
  */
 function ui_patterns_theme() {
-  /** @var \Drupal\ui_patterns\UiPatternBase $pattern */
+  /** @var \Drupal\ui_patterns\Plugin\PatternBase $pattern */
   $items = [
     'patterns_destination' => [
       'variables' => ['sources' => NULL, 'context' => NULL],
@@ -31,7 +31,7 @@ function ui_patterns_theme() {
  * Implements hook_library_info_build()
  */
 function ui_patterns_library_info_build() {
-  /** @var \Drupal\ui_patterns\UiPatternBase $pattern */
+  /** @var \Drupal\ui_patterns\Plugin\PatternBase $pattern */
 
   $definitions = [];
   foreach (UiPatterns::getManager()->getPatterns() as $pattern) {
